requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
supabase==2.7.4
openai==1.51.0
google-generativeai==0.8.3
anthropic==0.34.2
beautifulsoup4==4.12.3
requests==2.32.3
python-multipart==0.0.12
python-dotenv==1.0.1
Pillow==10.4.0
httpx==0.27.2
pydantic==2.9.2
aiohttp==3.10.10
